+++ srcs/SubChunk_faces.cpp
@@ -338,8 +338,8 @@ void SubChunk::processWestVertex(std::vector<Face> *faces, std::vector<int> *ver
 void SubChunk::addFace(ivec3 position, Direction dir, TextureType texture, bool isTransparent) {
     Face newFace;
     newFace.position = position;
-    newFace.size = vec2(0, 0);
-    // newFace.size = vec2(1, 1);
+    newFace.size = ivec2(0, 0);
+    // newFace.size = ivec2(1, 1);
     newFace.direction = dir;
     newFace.texture = texture;
     if (isTransparent)
